% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/networkfuncs.R
\name{convertD3}
\alias{convertD3}
\title{Return D3 direct Graph list}
\usage{
convertD3(g)
}
\arguments{
\item{g}{a graph object that you want to plot}
}
\value{
a list with numerical graph of links and a node list with names and groups
}
\description{
This function will take an igraph object and return a list that can be plotted as a D3network graph
}

