% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/networkfuncs.R
\name{mention_ext}
\alias{mention_ext}
\title{Extract mentions}
\usage{
mention_ext(txt)
}
\arguments{
\item{txt}{the text of your tweet}
}
\description{
extract all the mentions of other twitter users from a tweet
}

