% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/networkfuncs.R
\name{tweet_from_api}
\alias{tweet_from_api}
\title{tweets from API}
\usage{
tweet_from_api(tweetMin, dt)
}
\arguments{
\item{tweetMin}{minimum number of tweets a user id should have to be included in the network}

\item{dt}{the date you want data for, could be 'all' if you want all days}
}
\description{
Grab tweets from the API
}

