% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/networkfuncs.R
\name{tweet_subset}
\alias{tweet_subset}
\title{Subset tweets}
\usage{
tweet_subset(db, tweetMin = 0, dt = "all")
}
\arguments{
\item{db}{the name of the database that holds all the tweets, it should include the full path}

\item{tweetMin}{minimum number of tweets a user id should have to be included in the network}

\item{dt}{the date you want data for, could be 'all' if you want all days}
}
\description{
This will take user inputs and retrieve the data
}

